+++ twister2/master/src/java/edu/iu/dsc/tws/master/worker/JMWorkerController.java
@@ -9,7 +9,7 @@
 //  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 //  See the License for the specific language governing permissions and
 //  limitations under the License.
-package edu.iu.dsc.tws.master.client;
+package edu.iu.dsc.tws.master.worker;
 
 import java.net.InetAddress;
 import java.net.UnknownHostException;
